+++ app/__init__.py
@@ -1,3 +1,3 @@
 """Easy ASR Bench."""
 
-__version__ = "0.2.7"
+__version__ = "0.2.8"